IBX-4031: Isolated integration test
