+++ eZ/Publish/API/Repository/Tests/ContentServiceTest.php
@@ -6524,130 +6524,6 @@ class ContentServiceTest extends BaseContentServiceTest
         );
     }
 
-    public function testCopyNonTranslatableFieldsFromPublishedVersionToDraft(): void
-    {
-        $repository = $this->getRepository();
-        $permissionResolver = $repository->getPermissionResolver();
-        $contentTypeService = $repository->getContentTypeService();
-
-        $typeCreate = $contentTypeService->newContentTypeCreateStruct('nontranslatable');
-
-        $typeCreate->mainLanguageCode = 'eng-GB';
-        $typeCreate->remoteId = '1234567890abcdef';
-        $typeCreate->urlAliasSchema = '<title>';
-        $typeCreate->nameSchema = '<title>';
-        $typeCreate->names = [
-            'eng-GB' => 'Nontranslatable content type',
-        ];
-        $typeCreate->descriptions = [
-            'eng-GB' => '',
-        ];
-        $typeCreate->creatorId = $this->generateId(
-            'user',
-            $permissionResolver->getCurrentUserReference()->getUserId()
-        );
-        $typeCreate->creationDate = $this->createDateTime();
-
-        $titleFieldCreate = $contentTypeService->newFieldDefinitionCreateStruct('title', 'ezstring');
-        $titleFieldCreate->names = [
-            'eng-GB' => 'Title',
-        ];
-        $titleFieldCreate->descriptions = [
-            'eng-GB' => 'Title',
-        ];
-        $titleFieldCreate->fieldGroup = 'content';
-        $titleFieldCreate->position = 1;
-        $titleFieldCreate->isTranslatable = true;
-        $titleFieldCreate->isRequired = true;
-        $titleFieldCreate->isInfoCollector = false;
-        $titleFieldCreate->validatorConfiguration = [
-            'StringLengthValidator' => [
-                'minStringLength' => 0,
-                'maxStringLength' => 0,
-            ],
-        ];
-        $titleFieldCreate->fieldSettings = [];
-        $titleFieldCreate->isSearchable = true;
-        $titleFieldCreate->defaultValue = 'default title';
-
-        $typeCreate->addFieldDefinition($titleFieldCreate);
-
-        $bodyFieldCreate = $contentTypeService->newFieldDefinitionCreateStruct('body', 'ezstring');
-        $bodyFieldCreate->names = [
-            'eng-GB' => 'Body',
-        ];
-        $bodyFieldCreate->descriptions = [
-            'eng-GB' => 'Body',
-        ];
-        $bodyFieldCreate->fieldGroup = 'content';
-        $bodyFieldCreate->position = 2;
-        $bodyFieldCreate->isTranslatable = false;
-        $bodyFieldCreate->isRequired = false;
-        $bodyFieldCreate->isInfoCollector = false;
-        $bodyFieldCreate->validatorConfiguration = [
-            'StringLengthValidator' => [
-                'minStringLength' => 0,
-                'maxStringLength' => 0,
-            ],
-        ];
-        $bodyFieldCreate->fieldSettings = [];
-        $bodyFieldCreate->isSearchable = true;
-        $bodyFieldCreate->defaultValue = null;
-
-        $typeCreate->addFieldDefinition($bodyFieldCreate);
-
-        $contentTypeDraft = $contentTypeService->createContentType(
-            $typeCreate,
-            [$contentTypeService->loadContentTypeGroupByIdentifier('Media')],
-        );
-        $contentTypeService->publishContentTypeDraft($contentTypeDraft);
-
-        // Creating start content in eng-US language
-        $contentDraft = $this->createContentDraft(
-            'nontranslatable',
-            $this->generateId('location', 2),
-            [
-                'title' => 'Folder',
-            ]
-        );
-        $publishedContent = $this->contentService->publishVersion($contentDraft->getVersionInfo());
-
-        // Creating a draft in ger-DE language with the only field updated being 'title'
-        $gerDraft = $this->contentService->createContentDraft($publishedContent->contentInfo);
-
-        $contentUpdateStruct = new ContentUpdateStruct([
-            'initialLanguageCode' => self::GER_DE,
-            'fields' => $contentDraft->getFields(),
-        ]);
-
-        $contentUpdateStruct->setField('title', 'Folder GER', self::GER_DE);
-        $gerContent = $this->contentService->updateContent($gerDraft->getVersionInfo(), $contentUpdateStruct);
-
-        // Updating non-translatable field in eng-US language (allowed) and publishing it
-        $engContent = $this->contentService->createContentDraft($publishedContent->contentInfo);
-
-        $contentUpdateStruct = new ContentUpdateStruct([
-            'initialLanguageCode' => self::ENG_US,
-            'fields' => $contentDraft->getFields(),
-        ]);
-
-        $expectedBodyValue = 'Nontranslatable value';
-        $contentUpdateStruct->setField('title', 'Title v2', self::ENG_US);
-        $contentUpdateStruct->setField('body', $expectedBodyValue, self::ENG_US);
-
-        $engContent = $this->contentService->updateContent($engContent->getVersionInfo(), $contentUpdateStruct);
-        $this->contentService->publishVersion($engContent->getVersionInfo());
-
-        // Publishing ger-DE draft with the empty non-translatable field
-        $this->contentService->publishVersion($gerContent->getVersionInfo());
-
-        // Loading main content
-        $mainPublishedContent = $this->contentService->loadContent($engContent->id);
-        $bodyFieldValue = $mainPublishedContent->getField('body')->getValue();
-
-        self::assertSame($expectedBodyValue, $bodyFieldValue->text);
-    }
-
     /**
      * Create structure of parent folders with Locations to be used for Content hide/reveal tests.
      *
